fix Int check on large unsigned integers
Large unsigned integers can't always be converted to floats without
losing precision.  Checking the NOKp flag and then the NV flag will end
up checking against an inaccurate value and won't register as an
integer.

The `p` flags should not be used for type checking, as they represent
that any value exists in the related slot.  Their value may not be
accurate.

POK is still checked first, as values like '0e0' will set the IOK flag
if used as a number.